refactor(event-bus): leave only one event type for emitting
