+++ src/utils/EventBus/types/EventPayloadBase.ts
@@ -1,0 +1,17 @@
+import { EventAction } from './EventAction';
+import { Index as IndexType } from './indexing';
+
+/**
+ * Common fields for all events related to the document model
+ */
+export interface EventPayloadBase<Index extends IndexType, Action extends EventAction> {
+  /**
+   * The index of changed information
+   */
+  index: Index;
+
+  /**
+   * The action that was performed on the information
+   */
+  action: Action;
+}
