+++ src/utils/EventBus/types/EventTarget.d.ts
@@ -1,10 +1,10 @@
-import { DocumentModelEventMap } from './DocumentModelEventMap';
+import { EventMap } from './EventMap';
 
 /**
  * Augment EventTarget's addEventListener method to accept CustomEvent
  */
 declare global {
   interface EventTarget {
-    addEventListener<T extends keyof DocumentModelEventMap>(type: T, listener: (event: DocumentModelEventMap[T]) => void): void;
+    addEventListener<T extends keyof EventMap>(type: T, listener: (event: EventMap[T]) => void): void;
   }
 }
