+++ src/utils/EventBus/types/EventType.ts
@@ -1,19 +1,9 @@
 /**
- * Enumeration of the possible event types
+ * Event types enum that can be emitted by the document model
  */
 export enum EventType {
   /**
-   * Event indicating that some new information was added to the document model
+   * The document model has been changed
    */
-  Added = 'added',
-
-  /**
-   * Event indicating that some information was removed from the document model
-   */
-  Removed = 'removed',
-
-  /**
-   * Event indicating that some existing information was modified in the document model
-   */
-  Modified = 'modified',
+  CHANGED = 'changed',
 }
